Separate razer manager and razer layer
