+++ Project-Aurora/Project-Aurora/Settings/Layers/Control_RazerLayer.xaml.cs
@@ -56,28 +56,6 @@ namespace Aurora.Settings.Layers
         private void OnUserControlLoaded(object sender, RoutedEventArgs e)
         {
             SetSettings();
-
-            var version = RzHelper.GetSdkVersion();
-            var enabled = RzHelper.IsSdkEnabled();
-
-            SdkInstalledVersionValueLabel.Content = version.ToString();
-            if (!RzHelper.IsSdkVersionSupported(version))
-            {
-                SdkInstalledVersionValueLabel.Foreground = new SolidColorBrush(Colors.PaleVioletRed);
-                SdkInstallButton.Visibility = Visibility.Visible;
-            }
-            else
-            {
-                SdkInstalledVersionValueLabel.Foreground = new SolidColorBrush(Colors.LightGreen);
-                SdkInstallButton.Visibility = Visibility.Hidden;
-            }
-
-            SdkEnabledValueLabel.Content = enabled ? "Enabled" : "Disabled";
-            SdkEnabledValueLabel.Foreground = enabled ? new SolidColorBrush(Colors.LightGreen) : new SolidColorBrush(Colors.PaleVioletRed);
-
-            SdkConnectionStatusLabel.Content = Context.Loaded ? "Success" : "Failure";
-            SdkConnectionStatusLabel.Foreground = Context.Loaded ? new SolidColorBrush(Colors.LightGreen) : new SolidColorBrush(Colors.PaleVioletRed);
-
             Loaded -= OnUserControlLoaded;
         }
 
@@ -133,157 +111,5 @@ namespace Aurora.Settings.Layers
             SdkDumpToggleButton.IsEnabled = true;
             Context.StopDumpingData();
         }
-
-        private void OnSdkInstallButtonClick(object sender, RoutedEventArgs e)
-        {
-            SdkInstallButton.IsEnabled = false;
-
-            bool HandleErrorLevel(int errorlevel)
-            {
-                switch (errorlevel)
-                {
-                    case 3010:
-                        {
-                            MessageBox.Show("Razer SDK requested system restart!\nPlease reboot your pc and re-run the installation.",
-                                "Restart required!", MessageBoxButton.OK, MessageBoxImage.Exclamation);
-                            return false;
-                        }
-                }
-
-                return true;
-            }
-
-            void SetState(string name)
-                => Application.Current.Dispatcher.Invoke(() => SdkInstallButton.Content = name);
-
-            Task.Run(async () =>
-            {
-                try
-                {
-                    SetState("Uninstalling");
-                    var errorlevel = await UninstallAsync();
-                    if (!HandleErrorLevel(errorlevel))
-                        return false;
-
-                    SetState("Downloading");
-                    var path = await DownloadAsync();
-
-                    SetState("Installing");
-                    errorlevel = await InstallAsync(path);
-                    if (!HandleErrorLevel(errorlevel))
-                        return false;
-
-                    SetState("Done!");
-                }
-                catch (OperationCanceledException ex)
-                {
-                    SetState("Failure");
-                    MessageBox.Show($"{ex.Message}:\n{ex.InnerException.ToString()}",
-                        "Exception!", MessageBoxButton.OK, MessageBoxImage.Error);
-                    return false;
-                }
-
-                return true;
-            }).ContinueWith(t =>
-            {
-                if (t.Result)
-                {
-                    SetState("Success");
-                    MessageBox.Show("Installation successful!\nPlease restart Aurora for changes to take effect.",
-                        "Success!", MessageBoxButton.OK, MessageBoxImage.Information);
-                }
-                else
-                {
-                    SetState("Failure");
-                }
-            });
-        }
-
-        #region SDK Installer/Uninstaller helpers
-        private Task<int> UninstallAsync()
-        {
-            return Task.Run(() =>
-            {
-                if (RzHelper.IsSdkVersionSupported(RzHelper.GetSdkVersion()))
-                    return 0;
-
-                using (var hklm = RegistryKey.OpenBaseKey(RegistryHive.LocalMachine, RegistryView.Registry32))
-                {
-                    var key = hklm.OpenSubKey(@"Software\Razer Chroma SDK");
-                    var path = (string)key?.GetValue("UninstallPath", null);
-                    var filename = (string)key?.GetValue("UninstallFilename", null);
-
-                    if (path == null || filename == null)
-                        return 0;
-
-                    try
-                    {
-                        var processInfo = new ProcessStartInfo
-                        {
-                            FileName = filename,
-                            WorkingDirectory = path,
-                            Arguments = $"/S _?={path}",
-                            ErrorDialog = true
-                        };
-
-                        var process = Process.Start(processInfo);
-                        process.WaitForExit(120000);
-                        return process.ExitCode;
-                    }
-                    catch (Exception ex)
-                    {
-                        throw new OperationCanceledException("Razer SDK Uninstallation failed!", ex);
-                    }
-                }
-            });
-        }
-
-        private Task<string> DownloadAsync()
-        {
-            return Task.Run(() =>
-            {
-                var url = "http://cdn.razersynapse.com/156092369797u1UA8NRazerChromaBroadcasterSetup_v3.4.0630.061913.exe";
-
-                try
-                {
-                    using (var client = new WebClient())
-                    {
-                        var path = Path.ChangeExtension(Path.GetTempFileName(), ".exe");
-                        client.DownloadFile(url, path);
-                        return path;
-                    }
-                }
-                catch (Exception ex)
-                {
-                    throw new OperationCanceledException("Razer SDK Downloading failed!", ex);
-                }
-            });
-        }
-
-        private Task<int> InstallAsync(string installerPath)
-        {
-            return Task.Run(() =>
-            {
-                try
-                {
-                    var processInfo = new ProcessStartInfo
-                    {
-                        FileName = Path.GetFileName(installerPath),
-                        WorkingDirectory = Path.GetDirectoryName(installerPath),
-                        Arguments = "/S",
-                        ErrorDialog = true
-                    };
-
-                    var process = Process.Start(processInfo);
-                    process.WaitForExit(120000);
-                    return process.ExitCode;
-                }
-                catch (Exception ex)
-                {
-                    throw new OperationCanceledException("Razer SDK Installation failed!", ex);
-                }
-            });
-        }
-        #endregion
     }
 }
